requirements.txt
# DGA Qiyas Copilot - Python Dependencies

# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Configuration and Settings
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1

# Database
sqlalchemy==2.0.25

# Authentication
ldap3==2.9.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# Azure Providers
openai==1.10.0  # Azure OpenAI SDK
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
azure-core==1.29.6

# Google Providers
google-cloud-aiplatform==1.40.0
google-cloud-storage==2.14.0
google-cloud-discoveryengine==0.11.11
google-auth==2.26.2

# Utilities
tiktoken==0.5.2  # Token counting for OpenAI models
